requirements.txt
requests
pytest==4.4.1
websocket-client